requirements.txt
altair==5.4.1
annotated-types==0.7.0
argcomplete==3.5.0
attrs==24.2.0
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
DateTime==5.5
Faker==30.8.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.10.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pipx==1.7.1
platformdirs==4.3.6
protobuf==5.28.3
psycopg2==2.9.10
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
s3transfer==0.10.3
schedule==1.2.2
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.36
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
userpath==1.9.2
watchdog==5.0.3
zope.interface==7.1.1
